pyscripts/requirements.txt
Package              Version
-------------------- ---------
apipkg               2.1.0
appdirs              1.4.3
argon2-cffi          21.3.0
argon2-cffi-bindings 21.2.0
asn1crypto           0.24.0
bcrypt               4.0.1
Brlapi               0.8.3
certifi              2018.1.18
cffi                 1.15.1
chardet              3.0.4
construct            2.10.70
cryptography         40.0.2
cupshelpers          1.0
dataclasses          0.8
decorator            4.4.2
distlib              0.3.7
et-xmlfile           1.1.0
filelock             3.4.1
gitdb                4.0.9
GitPython            3.1.18
idna                 2.6
importlib-metadata   4.8.3
importlib-resources  5.4.0
iniconfig            0.0.0
isc                  2.0
louis                3.20.0
lxml                 4.9.3
nftables             0.1
openpyxl             3.1.2
ordered-set          4.0.2
packaging            21.3
paramiko             3.3.1
pip                  21.3.1
platformdirs         2.4.0
ply                  3.10
py                   1.10.0
pyasn1               0.4.2
pycairo              1.20.1
pycparser            2.21
pycryptodome         3.19.0
pycryptodomex        3.19.0
pycups               1.9.74
pycurl               7.43.0.2
PyGObject            3.42.2
pykeepass            4.0.6
PyNaCl               1.5.0
pyOpenSSL            21.0.0
pyparsing            2.4.7
pysmbc               1.0.23
PySocks              1.6.8
python-dateutil      2.8.2
requests             2.24.0
rpm                  4.14.3
setuptools           44.1.1
six                  1.16.0
slip                 0.6.5
slip.dbus            0.6.5
smmap                5.0.0
typing_extensions    4.1.1
urllib3              1.25.10
urlscan              0.9.10
urwid                2.0.1
virtualenv           20.17.1
zipp                 3.6.0
